product_inventory_drf/requirements.txt
Django==4.1.2
djangorestframework==3.15.1
psycopg2==2.9.1
Faker==25.3.0
django-extensions
django-debug-toolbar